+++ src/afterfact.rs
@@ -2060,6 +2060,7 @@ mod tests {
                 remove_duplicate_data: false,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             },
             geo_ip: None,
             output: Some(Path::new("./test_emit_csv.csv").to_path_buf()),
@@ -2151,6 +2152,7 @@ mod tests {
                 remove_duplicate_data: false,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             };
             let ch = mock_ch_filter
                 .get(&CompactString::from("security"))
@@ -2392,6 +2394,7 @@ mod tests {
                 remove_duplicate_data: false,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             },
             geo_ip: None,
             output: Some(Path::new("./test_emit_csv_multiline.csv").to_path_buf()),
@@ -2485,6 +2488,7 @@ mod tests {
                 remove_duplicate_data: false,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             };
             let ch = mock_ch_filter
                 .get(&CompactString::from("security"))
@@ -2712,6 +2716,7 @@ mod tests {
                 remove_duplicate_data: true,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             },
             geo_ip: None,
             output: Some(Path::new("./test_emit_csv_remove_duplicate.csv").to_path_buf()),
@@ -2803,6 +2808,7 @@ mod tests {
                 remove_duplicate_data: false,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             };
             let ch = mock_ch_filter
                 .get(&CompactString::from("security"))
@@ -3041,6 +3047,7 @@ mod tests {
                 remove_duplicate_data: true,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             },
             geo_ip: None,
             output: Some(Path::new("./test_emit_csv_remove_duplicate.json").to_path_buf()),
@@ -3132,6 +3139,7 @@ mod tests {
                 remove_duplicate_data: true,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             };
             let ch = mock_ch_filter
                 .get(&CompactString::from("security"))
@@ -3218,7 +3226,7 @@ mod tests {
             println!("message: {detect_infos:?}");
         }
 
-        let expect_target = vec![
+        let expect_target = [
             vec![
                 (
                     "Timestamp",
@@ -3444,6 +3452,7 @@ mod tests {
                 remove_duplicate_data: true,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             },
             geo_ip: None,
             output: Some(Path::new("./test_multiple_data_in_details.json").to_path_buf()),
@@ -3536,6 +3545,7 @@ mod tests {
                 remove_duplicate_data: true,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             };
             let ch = mock_ch_filter
                 .get(&CompactString::from("security"))
@@ -3597,8 +3607,7 @@ mod tests {
                 Profile::Computer(test_computername.into());
         }
 
-        let expect_target = vec![
-            vec![
+        let expect_target = [vec![
                 (
                     "Timestamp",
                     CompactString::from(
@@ -3652,8 +3661,7 @@ mod tests {
                     "Tags",
                     CompactString::from("[\n        \"".to_string() + test_attack + "\"\n    ]"),
                 ),
-            ]
-        ];
+            ]];
         let mut expect_str = String::default();
         for (target_idx, target) in expect_target.iter().enumerate() {
             let mut expect_json = "{\n".to_string();
@@ -3792,6 +3800,7 @@ mod tests {
                 remove_duplicate_data: false,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             },
             geo_ip: None,
             output: Some(Path::new("./test_emit_csv_json.json").to_path_buf()),
@@ -3882,6 +3891,7 @@ mod tests {
                 remove_duplicate_data: false,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             };
             let ch = mock_ch_filter
                 .get(&CompactString::from("security"))
@@ -4061,6 +4071,7 @@ mod tests {
                 remove_duplicate_data: false,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             },
             geo_ip: None,
             output: Some(Path::new("./test_emit_csv_jsonl.jsonl").to_path_buf()),
@@ -4151,6 +4162,7 @@ mod tests {
                 remove_duplicate_data: false,
                 remove_duplicate_detections: false,
                 no_wizard: true,
+                include_status: None,
             };
             let ch = mock_ch_filter
                 .get(&CompactString::from("security"))